+++ src/field/float_field.h
@@ -8,7 +8,7 @@ class FloatField : public Field {
 public:
   FloatField() = default;
 
-  explicit FloatField(double data);
+  explicit FloatField(float data);
 
   explicit FloatField(const uint8_t *src);
 
@@ -37,7 +37,7 @@ public:
   friend std::ostream &operator<<(std::ostream &os, const FloatField &field);
 
 private:
-  double _data{};
+  float _data{};
 };
 
 } // namespace Neru
